requirements.txt
snowflake-connector-python==3.12.3
python-dotenv==1.0.1
pandas==2.2.3
pymupdf==1.25.1
scikit-learn==1.6.1
langchain>=0.3.0
langchain-chroma>=0.1.0
langchain-core>=0.3.0
langchain-text-splitters>=0.3.0
chromadb>=0.5.0
python-docx>=1.1.0
